Reduce hand class concerns
The hand class was originally built to fully describe a hand,
including how one hand compared to other hands, and what it represented
in a game of poker. The class had too many responsibilities.

All high-level evaluation of the hand was removed and put into an
evaluator module. Now, hands only know the cards they contain, and the
"answers" to questions about their identity that have nothing to do
with how poker is scored. For example, a hand can answer the question "are all your
cards the same suit?" if the flush? method is called.