Categorize identification logic
Identifying a hand differs based on the type of poker being played.
In this case, the identification is based on the rules of texas hold em.

Breaking that logic out into its own file means it is separated from
hand evaluation, and it can be stored with other roles that follow
texas hold em rules.