Create single call method for service objects
Each service object should perform only one task and contain only one
public method apart from initialize(). This method should perform a
task in line with the name of the class. Since the name of the class
describes the task that is being performed, we can change the name of
the public method to "call". This makes calls to the service object
shorter and easier to scan.