Use individual set searches
Now that we have objects that search for individual types of sets, we
can use those instead of fetching all sets and filtering by type.

This change is important. Without it, since different sets have their
own classes, we would have to filter the fetched list of all sets by
class name. This way, we just pass the hand to the appropriate set
searcher and it returns all sets of the type we were looking for,
without us knowing how to tell the difference between types.