Wrap hand value hash in object
Instead of storing the hand value mapping in a global constant, we can
wrap it in an accessor object to abstract away the implementation of
the hand to value mapping (in this case it's a hash, but it could change
in the future). This way, client code that needs to retrieve a hand value
is not aware of the underlying data structure, or any additional tasks
that occur in the process of mapping a hand type to a value.