Inject hand value mapping dependency
Since we have the hand values stored in an object, we can pass them
to the hand type comparison object when it is initialized. This means
the comparison object is not tied to texas hold em rankings and
can be more easily reused.

In the future, more work should be done to remove the texas hold em
identify object from the comparison class, so that it is completely
separated from any specific style of poker.